reformatted JavaConcepts
