+++ javaparser-core-testing/src/test/test_sourcecode/com/github/javaparser/printer/JavaConceptsUgly.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright (C) 2007-2010 Júlio Vilmar Gesser.
+ * Copyright (C) 2011, 2013-2016 The JavaParser Team.
+ *
+ * This file is part of JavaParser.
+ *
+ * JavaParser can be used either under the terms of
+ * a) the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ * b) the terms of the Apache License
+ *
+ * You should have received a copy of both licenses in LICENCE.LGPL and
+ * LICENCE.APACHE. Please refer to those files for details.
+ *
+ * JavaParser is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ */
+package japa.bdd.samples;
+
+@Deprecated
+public class JavaConceptsUgly {
+
+    static int x = 0;
+
+    public static void main(String[] args) {
+        x = +x;
+        x = ~x;
+        --x;
+        boolean b = !false;
+        x &= 2;
+        x |= 2;
+        x ^= 2;
+        x -= 2;
+        x %= 2;
+        x /= 2;
+        x *= 2;
+        x <<= 2;
+        x >>= 2;
+        x >>>= 2;
+        b = b || false;
+        b = b | false;
+        b = b & false;
+        b = b ^ false;
+        b = b != false;
+        b = x > 1;
+        b = x < 1;
+        b = x >= 1;
+        b = x <= 1;
+        x = x << 1;
+        x = x >> 1;
+        x = x >>> 1;
+        x = x - 1;
+        x = x * 1;
+        x = x % 1;
+        x = x / 1;
+    }
+}
